tag 8.2.22 with fixes for OSI file changes

+++ UltimaSDK/Ultima/TileData.cs
@@ -284,8 +284,9 @@ namespace Ultima
         /// </summary>
         public TileFlag Flags
         {
-            get { 
-                return m_Flags; 
+            get
+            {
+                return m_Flags;
             }
             set { m_Flags = value; }
         }
@@ -908,7 +909,7 @@ namespace Ultima
 
                         bin.Write((int)m_LandData[i].Flags);
                         if (useNewTileDataFormat)
-                            bin.Write((int)m_LandData[i].Unk1);
+                            bin.Write(m_LandData[i].Unk1);
 
                         bin.Write(m_LandData[i].TextureID);
                         byte[] b = new byte[20];
@@ -929,7 +930,7 @@ namespace Ultima
 
                         bin.Write((int)m_ItemData[i].Flags);
                         if (useNewTileDataFormat)
-                            bin.Write((int)m_ItemData[i].Unk1);
+                            bin.Write(m_ItemData[i].Unk1);
 
                         bin.Write(m_ItemData[i].Weight);
                         bin.Write(m_ItemData[i].Quality);
